--- Example/demo.wiretapp/demo.wiretapp.xcodeproj/project.xcworkspace/xcshareddata/swiftpm/Package.resolved
@@ -1,16 +1,0 @@
-{
-  "object": {
-    "pins": [
-      {
-        "package": "wiretapp",
-        "repositoryURL": "https://github.com/MrMUsmanAli/wiretapp",
-        "state": {
-          "branch": null,
-          "revision": "fd075784ad946022da57ce4a05879cbbde64432f",
-          "version": null
-        }
-      }
-    ]
-  },
-  "version": 1
-}
