+++ Package.swift
@@ -8,13 +8,9 @@ let package = Package(
     products: [
         // Products define the executables and libraries a package produces, and make them visible to other packages.
         .library(
-            name: "wiretappcore",
-            targets: ["wiretappcore"]
-        ),
-        .library(
-            name: "wiretapptest",
-            targets: ["wiretapptest"]
-        ),
+            name: "wiretapp",
+            targets: ["wiretapp"]
+        )
     ],
     dependencies: [
         // Dependencies declare other packages that this package depends on.
@@ -24,20 +20,14 @@ let package = Package(
         // Targets are the basic building blocks of a package. A target can define a module or a test suite.
         // Targets can depend on other targets in this package, and on products in packages this package depends on.
         .target(
-            name: "wiretappcore",
+            name: "wiretapp",
             dependencies: [],
             path: "wiretapp",
             exclude: [ "Example/" ]
         ),
-        .target(
-            name: "wiretapptest",
-            dependencies: ["wiretappcore"],
-            path: "wiretapptest",
-            exclude: [ "Example/" ]
-        ),
         .testTarget(
             name: "wiretappTests",
-            dependencies: ["wiretappcore"],
+            dependencies: ["wiretapp"],
             exclude: [ "Example/" ]
         ),
     ]
